feat(guest-agent): add api_to_cli_init telemetry metric
Record E2E time from API run creation to the CLI emitting its first
init event (system/init for Claude Code, thread.started for Codex).
Extract the shared E2E timing logic into a new `timing` module.

Closes #3244
